git-rr: sort/grant FILENNNNN names by usage
i.e. ensure FILE00001 is the most-used / "first" in ASCIIbetical
order, rather than the very first file name found in the list.
This ensures one can pick a neat set of recognisable colours for
each FILENNNNN and the file that is used the most will be the most
distinguishable in the list.