[git-files-changed-from] fix invocation and sorting
There's a difference when the command is invoked standalone and when
it's invoked via a Git alias, i.e. between:

    $ git-files-changed-from
    $ git fcom

... having the following in ~/.gitconfig:

    [alias]
    fcom = !git-files-changed-from "origin/$(git main)"

When invoked standalone, it gets the current PWD and no GIT_PREFIX.

When invoked through a Git alias, the way I use it most often, the PWD
is the repo's root, and the GIT_PREFIX is set to the remainder path that
you're "really" launching the command from.

My previous change only made things work via the alias, but not via the
fuller command invocation.
The setting of GIT_PREFIX and the subsequent "cd" command harmonise
those two invocation methods by ensuring:
- the command is _always_ ran from the repo's root
- the command has a proper GIT_PREFIX set (if not already set)

After using this as it was for a bit, I grew frustrated that while in a
subdirectory the files modified inside that subdirectory "possibly
sorted after" files modified in another directory's path. I've added a
Perl invocation to ensure:
- files without a "../" prefix "sort before, but still asciibetically"
  than files with a "../" prefix
- anything else (when both have a "../" prefix, or when neither does)
  sorts asciibetically as before.

As such, my "CTRL+F" when done in a subdirectory _properly_ shows the
files modified in that subdirectory (hence with no relative "../" path)
*before* files modified wherever else.